requirements.txt
click==8.1.3
colorama==0.4.4
Faker==13.11.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.2
pymongo==4.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
